fix: cmdline autocompletion
cmp.setup.cmdline was missing from the config so it wasn't doing
anything. Oops. And even if that was set Noice was using nui and
wildmenu for cmdline autocompletion. Double oops. And it didn't work
correctly since some of the cmp mappings were overriding default keymaps
for wildmenu. Triple oops.

Also set Up and Down in cmdline completion to behave like they do in
rest of the cmp menus.