chore(ftplugin): get rid of trivial ftplugins
Files that only set indentation type and width should be handled by
editorconfig. If not present in the repo, it'll eventually fall back to
the one in home dir.

Still left the helm plugin there because neovim can identify files as
helm files (with the right lsp) even when their file extensions aren't
special in any way.