feat(vimade): try ncmode="buffers"
material.nvim still sets the background of unfocused buffers to be
darker than the focused one which highlights which of the duplicate
buffers is currently active while making other buffers way darker. Kinda
nice with a short testing, let's see how it turns out when used for a
bit longer.