feat(plugins): add work-time-calculator
Started work on a script that calculates my working hours from my daily
notes and quickly realized I want to put it into a separate plugin.

Using my own fork and a local dev branch of the time-calculator
dependency that includes the necessary changes to mave it work for now.
Need to get around updating the readme about them and making pull
requests.