feat(obsidian)!: disable checkbox toggling with enter
If it would work so that if you hit enter on a line that already has a
checkbox it would only then toggle it it would be awesome. But as it
always adds checkbox to the line if it already didn't have one it just
gets in the way.

Disabled it, and added a keyboard shortcut for toggling checkbox. Also
removed the unnecessary ObsidianFollowLink binding as `gf` does the
trick.