feat(blink): fully customize bindings
C-n and C-p will now show the menu if it isn't visible already. C-n also
inserts the first item. I wish C-p would insert the last item but
could't get it working so `¯\_(ツ)_/¯`

Since C-n and C-p double as showing the menu as well, C-Space is now
disabled as redundant.

Esc will now cancel the completion. As a result, C-e is disabled as
redundant.